requirements.txt
fastapi
uvicorn[standard]
pydantic
qdrant-client
python-dotenv
requests
streamlit
PyPDF2
